Added support for the "final" keyword to define a variable.
Updated the tests so they all pass.
Updated the library as required so the new elements added while working on the compiler are all tested (100% coverage again).
Removed the wrong duplicate of the CMAKE_BUILD_TYPE from the coverage script.
Fixed the doxygen setup to get indices in alphabetical order.

+++ as2js/tests/test_as2js_rc.cpp
@@ -665,7 +665,14 @@ void As2JsRCUnitTests::test_load_from_user_config()
     // not reach here if the .rc already existed!)
     as2js::String config(home);
     config += "/.config";
-    CPPUNIT_ASSERT(mkdir(config.to_utf8().c_str(), 0700) == 0); // usually this is 0755, but for security we cannot take that risk...
+    std::cout << " --- config path \"" << config << "\" --- ";
+    bool del_config(true);
+    if(mkdir(config.to_utf8().c_str(), 0700) != 0) // usually this is 0755, but for security we cannot take that risk...
+    {
+        // if this mkdir() fails, it is because it exists
+        CPPUNIT_ASSERT(errno == EEXIST);
+        del_config = false;
+    }
     as2js::String as2js_conf(config);
     as2js_conf += "/as2js";
     CPPUNIT_ASSERT(mkdir(as2js_conf.to_utf8().c_str(), 0700) == 0);
@@ -826,7 +833,10 @@ void As2JsRCUnitTests::test_load_from_user_config()
 
     // if possible get rid of the directories (don't check for errors)
     rmdir(as2js_conf.to_utf8().c_str());
-    rmdir(config.to_utf8().c_str());
+    if(del_config)
+    {
+        rmdir(config.to_utf8().c_str());
+    }
 }
 
 
@@ -1038,7 +1048,14 @@ void As2JsRCUnitTests::test_empty_home()
 
     as2js::String config(home);
     config += "/.config";
-    CPPUNIT_ASSERT(mkdir(config.to_utf8().c_str(), 0700) == 0);
+    std::cout << " --- config path \"" << config << "\" --- ";
+    bool del_config(true);
+    if(mkdir(config.to_utf8().c_str(), 0700) != 0) // usually this is 0755, but for security we cannot take that risk...
+    {
+        // if this mkdir() fails, it is because it exists
+        CPPUNIT_ASSERT(errno == EEXIST);
+        del_config = false;
+    }
 
     as2js::String rc_path(config);
     rc_path += "/as2js";
